fixed a bug in delete task file in server side

+++ server/routes/deleteTask.js
@@ -5,10 +5,14 @@ const authenticateToken = require("../middlewares/authenticateToken");
 
 router.put("/", authenticateToken, async (req, res) => {
   try {
-    const query = 'DELETE FROM `db_tasks` WHERE `task_id` = ?';
-    const value = [req.body.taskId];
-    await database(query, value);
-    return(res.status(200).json("Successfully deleted the task."));
+    if (req.body.userData.signedIn === true) {
+      const query = 'DELETE FROM `db_tasks` WHERE `task_id` = ?';
+      const value = [req.body.taskId];
+      await database(query, value);
+      return(res.status(200).json("Successfully deleted the task."));
+    } else {
+      return(res.status(400).json({error: "Please sign in."}));
+    };
   } catch (error) {
     return(res.status(400).json(error));
   };
